domain/utils: move password autogen
