+++ dexios-domain/src/utils.rs
@@ -1,5 +1,7 @@
 // TODO(pleshevskiy): dedup these utils
 
+use dexios_core::protected::Protected;
+
 #[cfg(test)]
 mod test {
     use dexios_core::primitives::{get_nonce_len, Algorithm, Mode, MASTER_KEY_LEN, SALT_LEN};
@@ -28,6 +30,41 @@ mod test {
         StdRng::seed_from_u64(MASTER_KEY_SEED).fill_bytes(&mut master_key);
         Protected::new(master_key)
     }
+}
+
+// this autogenerates a passphrase, which can be selected with `--auto`
+// it reads the EFF large list of words, and puts them all into a vec
+// 3 words are then chosen at random, and 6 digits are also
+// the 3 words and the digits are separated with -
+// the words are also capitalised
+// this passphrase should provide adequate protection, while not being too hard to remember
+pub fn gen_passphrase() -> Protected<String> {
+    use rand::{prelude::StdRng, Rng, SeedableRng};
+    let collection = include_str!("wordlist.lst");
+    let words = collection.lines().collect::<Vec<_>>();
+
+    let mut passphrase = String::new();
+
+    for _ in 0..3 {
+        let index = StdRng::from_entropy().gen_range(0..=words.len());
+        let word = words[index];
+        let capitalized_word = word
+            .char_indices()
+            .map(|(i, ch)| match i {
+                0 => ch.to_ascii_uppercase(),
+                _ => ch,
+            })
+            .collect::<String>();
+        passphrase.push_str(&capitalized_word);
+        passphrase.push('-');
+    }
+
+    for _ in 0..6 {
+        let number: i64 = StdRng::from_entropy().gen_range(0..=9);
+        passphrase.push_str(&number.to_string());
+    }
+
+    Protected::new(passphrase)
 }
 
 #[cfg(test)]
