+++ dexios/src/global/key.rs
@@ -1,7 +1,6 @@
 use anyhow::{Context, Result};
 use dexios_core::protected::Protected;
 use dexios_core::Zeroize;
-use rand::{prelude::StdRng, Rng, SeedableRng};
 
 use crate::{global::states::PasswordState, warn};
 
@@ -27,38 +26,4 @@ pub fn get_password(pass_state: &PasswordState) -> Result<Protected<Vec<u8>>> {
             warn!("The passwords aren't the same, please try again.");
         }
     })
-}
-
-// this autogenerates a passphrase, which can be selected with `--auto`
-// it reads the EFF large list of words, and puts them all into a vec
-// 3 words are then chosen at random, and 6 digits are also
-// the 3 words and the digits are separated with -
-// the words are also capitalised
-// this passphrase should provide adequate protection, while not being too hard to remember
-pub fn generate_passphrase() -> Protected<String> {
-    let collection = include_str!("wordlist.lst");
-    let words = collection.lines().collect::<Vec<_>>();
-
-    let mut passphrase = String::new();
-
-    for _ in 0..3 {
-        let index = StdRng::from_entropy().gen_range(0..=words.len());
-        let word = words[index];
-        let capitalized_word = word
-            .char_indices()
-            .map(|(i, ch)| match i {
-                0 => ch.to_ascii_uppercase(),
-                _ => ch,
-            })
-            .collect::<String>();
-        passphrase.push_str(&capitalized_word);
-        passphrase.push('-');
-    }
-
-    for _ in 0..6 {
-        let number: i64 = StdRng::from_entropy().gen_range(0..=9);
-        passphrase.push_str(&number.to_string());
-    }
-
-    Protected::new(passphrase)
 }
